change database validation
database wrapper now validates individual rows instead of validating the
type of the entire array of rows. It turns out that validating
the properties of the array as a whole should be left to the callsite
because the callsite will typically want to control what error is thrown
(e.g. specific 404 errors when the array is empty);