improve category filter behavior
category filter controls will not be shown when there are no tasks,
because the "show all" button on its own would appear to not do anything
and would be confusing to users.

the edit task dialog now prevents changing the task category to
"show all" since this would cause weird behavior of the category filter'
controls.