bump augustus to 3.1.0
augustus 3.1.0 added better support for literal unions and unions of
more than two schemas, which allowed refactoring some ugly code that
previously required nested unions of schemas. this new feature is only
used in the server, but i bumped the versions used by the app and core
packages as well for consistency.