fix server wrapper to send response representation
the server endpoint handler wrapper previously sent the result of
the handler function (which is of type Response) as the response
instead of using the endpoint response schema to convert to a ResRepr.
This distinction doesn't matter in Busybody because all the API response
types in have the same domain and representation types, but I'm
correcting so that it will work even if Response and ResRepr are
different in case I reuse this code in the future.